+++ libklio/types.hpp
@@ -47,7 +47,7 @@ namespace klio {
   // "tables" of sensor data. 
   typedef boost::multi_array<double, 2> sensordata_array_t;
   typedef sensordata_array_t::index sensordata_array_idx_t;
-  typedef boost::tuple<sensordata_array_t, int, int> sensordata_table_t;
+  typedef boost::tuple<sensordata_array_t, unsigned long int, unsigned long int> sensordata_table_t;
 
 }
